rebuild 1.55.0 boost libs and add libboost_log_setup
The setup lib is needed when specifying and sort of log formatter. Strangely libboost_log is significantly smaller than the previous build, who knows what was in the old version, maybe it statically linked other stuff in?

Boost 1.55.0 libs were built on OS X 10.9.4 with Xcode 5.1.1 generally following the Cinder instructions:
  http://libcinder.org/docs/v0.8.5/_cinder_boost.html
and replacing the two boost/atomic files mentioned here:
  https://trac.macports.org/ticket/42282

And the build command used was:
  ./bootstrap.sh –with-libraries=log,thread && ./b2 -a -j8 toolset=clang-osx link=static stage